Added code for set-amount commands & edited incomming command parsing

+++ txserver/event/schema.go
@@ -16,16 +16,16 @@ type Event struct {
 // UnmarshalBSONValue is an implementation that helps in decoding MongoDB bson response to golang struct
 func (e *Event) UnmarshalBSONValue(t bsontype.Type, data []byte) error {
 	var rawData bson.Raw
-	
+
 	err := bson.Unmarshal(data, &rawData)
 	if err != nil {
-		log.Printf("Error unmarshalling bytes to type bson.RAW: %s, error: %s", string(data),  err)
+		log.Printf("Error unmarshalling bytes to type bson.RAW: %s, error: %s", string(data), err)
 		return err
 	}
 
 	err = rawData.Lookup("eventType").Unmarshal(&e.EventType)
 	if err != nil {
-		log.Printf("Error unmarshalling eventType from rawBson: %+v, error: %s", rawData,  err)
+		log.Printf("Error unmarshalling eventType from rawBson: %+v, error: %s", rawData, err)
 	}
 
 	switch e.EventType {
@@ -50,13 +50,13 @@ func (e *Event) UnmarshalBSONValue(t bsontype.Type, data []byte) error {
 	}
 
 	err = rawData.Lookup("data").Unmarshal(e.Data)
-	if err != nil { 
+	if err != nil {
 		log.Printf("Couldn't Marshal rawBson : %+v, got error: %s", rawData, err)
 		return err
 	}
 
 	return nil
-}	
+}
 
 // UserCommand: Any command issued by the user
 type UserCommand struct {
@@ -67,7 +67,7 @@ type UserCommand struct {
 	Username       string  `xml:"username"`
 	StockSymbol    string  `xml:"stockSymbol"`
 	Filename       string  `xml:"filename"`
-	Funds          float64 `xml:"funds"`
+	Funds          float32 `xml:"funds"`
 }
 
 // QuoteServer: Any communication with the quoter server
@@ -75,7 +75,7 @@ type QuoteServer struct {
 	Timestamp       int     `xml:"timestamp"`
 	Server          string  `xml:"server"`
 	TransactionNum  int     `xml:"transactionNum"`
-	Price           float64 `xml:"price"`
+	Price           float32 `xml:"price"`
 	StockSymbol     string  `xml:"stockSymbol"`
 	Username        string  `xml:"username"`
 	QuoteServerTime int     `xml:"quoteServerTime"`
@@ -89,7 +89,7 @@ type AccountTransaction struct {
 	TransactionNum int     `xml:"transactionNum"`
 	Action         string  `xml:"action"`
 	Username       string  `xml:"username"`
-	Funds          float64 `xml:"funds"`
+	Funds          float32 `xml:"funds"`
 }
 
 // SystemEvent: Any event that is triggered by our system. For example, buying a stock because a trigger was set by the user.
@@ -101,7 +101,7 @@ type SystemEvent struct {
 	Username       string  `xml:"username"`
 	StockSymbol    string  `xml:"stockSymbol"`
 	Filename       string  `xml:"filename"`
-	Funds          float64 `xml:"funds"`
+	Funds          float32 `xml:"funds"`
 }
 
 // ErrorEvent: Any error that occurs for a transaction with the quote server
@@ -113,7 +113,7 @@ type ErrorEvent struct {
 	Username       string  `xml:"username"`
 	StockSymbol    string  `xml:"stockSymbol"`
 	Filename       string  `xml:"filename"`
-	Funds          float64 `xml:"funds"`
+	Funds          float32 `xml:"funds"`
 	ErrorMessage   string  `xml:"errorMessage"`
 }
 
@@ -126,6 +126,6 @@ type Debug struct {
 	Username       string  `xml:"username"`
 	StockSymbol    string  `xml:"stockSymbol"`
 	Filename       string  `xml:"filename"`
-	Funds          float64 `xml:"funds"`
+	Funds          float32 `xml:"funds"`
 	DebugMessage   string  `xml:"debugMessage"`
 }
